% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/heidi.R
\name{callAPI}
\alias{callAPI}
\title{Call HEIDI API}
\usage{
callAPI(passURL)
}
\arguments{
\item{passURL}{A string containing a HEIDI URI and parameters, usually assembled by other \pkg{heidi} functions.}
}
\description{
\code{callAPI} passes the supplied URL to HEIDI, authenticating using the global environment variable \code{base64userpass}, which should have already been assigned a value from \code{\link{setup}}. This function is used by the other \pkg{heidi} functions and should not normally be called directly by a user.
}
\keyword{internal}

