% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/heidi.R
\name{GetData}
\alias{GetData}
\title{Get HEIDI data}
\usage{
GetData(rowtype, year, valuetype, field1 = "", field2 = "", field3 = "",
  field4 = "", field5 = "")
}
\arguments{
\item{rowtype}{Supply a row type identifier specify the row type required. Must be compatible with the other arguments.}

\item{year}{Supply a year identifier to retrieve the value type for this year.}

\item{valuetype}{Supply a value type identifier to specify the measure required. Must be compatible with the other arguments.}

\item{field1,field2,field3,field4,field5}{Optionally supply up to five fields with which to break down the value type. Must be compatible with the other arguments. \strong{A field id for a particular field is not consistent across years}.}
}
\value{
Returns a dataframe containing the specified data. The data is in 'long' rather than 'wide' format, i.e. with one row per rowtype/field combination.
}
\description{
\code{GetData} retrieves the specified dataset from HEIDI. Ids for row type, year and value type are used to specify the required data, with up to five fields (field1, field 2, etc.) being provided to break the data down into further dimensions.
}
\details{
\code{GetData} retrieves data for the specified \code{year} and \code{valuetype} (e.g. 2011/12 Student FTEs). The value type (which could be called a "measure") is then broken down into more detail by the \code{rowtype} and the optionally-supplied \code{fields}.

Most of the functions in the \pkg{heidi} package exist to provide a means to query HEIDI and find out which datasets are available and to ascertain the specific ids of rowtypes, valuetypes, years and fields to feed to the GetData function.

Note that the results are rounded by HEIDI (often to the nearest 5) so providing too many \code{fields} will increase the rounding error as the \code{valuetype} is broken down into smaller and smaller parts.
}
\examples{
\dontrun{
# Get HE students FTEs for each institution for 2012/13 broken down by course year
heidi_data <- GetData(rowtype = '3297', year = '68682', valuetype = '3517-3508', field1 = '70469')
}
}

