% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/heidi.R
\name{GetFieldXML}
\alias{GetFieldXML}
\title{Get field info (XML)}
\usage{
GetFieldXML(fieldID)
}
\description{
\code{GetField} fetches additional information about a given field and returns the result in the raw XML form.
}
\keyword{internal}

