% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/heidi.R
\name{ListDomains}
\alias{ListDomains}
\title{List available Domains}
\usage{
ListDomains()
}
\description{
\code{ListDomains} fetches the names and identifiers of the available domains and returns them in a dataframe.
}
\details{
Domains are just a categorisation of Value Types to allow filtering of the hundreds of Value Types. Examples are: HESA Student, HESA Staff.
}
\examples{
\dontrun{
Domains <- ListDomains()
}
}

