% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/heidi.R
\name{ListFields}
\alias{ListFields}
\title{List available Fields}
\usage{
ListFields(rowtype = "", year = "", domain = "", valuetype = "")
}
\arguments{
\item{rowtype}{Supply a row type identifier to filter the list of available fields.}

\item{year}{Supply a year identifier to filter the list of available fields.}

\item{domain}{Supply a domain identifier to filter the list of available fields. Not necessary if a value type is provided.}

\item{valuetype}{Supply a value type identifier to filter the list of available fields.}
}
\description{
\code{ListFields} fetches the names and identifiers of the available fields and returns them in a dataframe. Ordinarily, ids for row type, year and value type/domain are provided to filter the list of fields.
}
\details{
A Field is dimension that breaks down a value type. For example, the value type 'HE students Full-time equivalent' could be broken down by the fields Subject, Gender, Nationality, Mode of Study.
}
\examples{
\dontrun{
Fields <- ListFields(rowtype = '3297', year = '68682', valuetype = '3517-3508')
}
}

