% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/heidi.R
\name{ListRowTypes}
\alias{ListRowTypes}
\title{List available Row Types}
\usage{
ListRowTypes(year = "")
}
\arguments{
\item{year}{Optionally supply a year identifier to filter the list of available row types for that year.}
}
\description{
\code{ListRowTypes} fetches the names and identifiers of the available row types and returns them in a dataframe.
}
\details{
A Row Type is a key comparison dimension that occurs in many underlying tables. Examples are: Institution, Cost Centre.
}
\examples{
\dontrun{
RowTypes <- ListRowTypes(year = '68682')  # 'year' refers to the year id
}
}

