% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/heidi.R
\name{ListValueTypes}
\alias{ListValueTypes}
\title{List available Value Types}
\usage{
ListValueTypes(rowtype = "", year = "", domain = "")
}
\arguments{
\item{rowtype}{Supply a row type identifier to filter the list of available value types.}

\item{year}{Supply a year identifier to filter the list of available value types.}

\item{domain}{Supply a domain identifier to filter the list of available value types.}
}
\description{
\code{ListValueTypes} fetches the names and identifiers of the available value types and returns them in a dataframe. Ordinarily, ids for row type, year and domain are provided to filter the list of value types.
}
\details{
A Value Type is a measure. Examples are: Headcount, Full Time Equivalents, UCAS Accepted Applicants.
}
\examples{
\dontrun{
ValueTypes <- ListValueTypes(rowtype = '3297', domain = '3308')
# Note, 'year' is unspecified to retrieve value types for all years
# applicable to the specified rowtype and domain
}
}

