% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/heidi.R
\name{ListYears}
\alias{ListYears}
\title{List available Years}
\usage{
ListYears(rowtype = "")
}
\arguments{
\item{rowtype}{Optionally supply a row type identifier to filter the list of available years for that row type.}
}
\description{
\code{ListYears} fetches the names and identifiers of the available years and returns them in a dataframe.
}
\details{
A Year is an academic year. Examples are: 2008/09, 2009/10.
}
\examples{
\dontrun{
Years <- ListYears(rowtype = '3297')  # 'rowtype' refers to the row type id
}
}

