% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/heidi.R
\name{setup}
\alias{setup}
\title{Setup HEIDI authentication}
\usage{
setup(userid, password)
}
\arguments{
\item{userid}{Your 4-digit heidi organisation identifier. If you don't know it it is available from the HEIDI local administrator web interface. It appears under "Manage API Keys" once a key has been created within the web interface, under the "Organisation" field.}

\item{password}{This is the "API Key" specific to your organisation. The HEIDI local administrator may need to create an API key from within the HEIDI web interface.}
}
\description{
\code{setup} converts the supplied heidi organisation identifier (\code{userid}) and API key (\code{password}) into its base64 equivalent, which is then used to authenticate subsequent API calls. The returned value must be saved into a variable called "base64userpass" which is accessed by other functions in the \pkg{heidi} package.
}
\examples{
base64userpass <- setup('1234', '1f1f1f1f-2e2e-3d3d-4c4c-5b5b5b5b5b5b')
}

